model: autosklearn: Updated to support archive storage
