+++ model/autosklearn/dffml_model_autosklearn/autoclassifier.py
@@ -1,43 +1,28 @@
 import autosklearn.classification
 
-from dffml.model.model import Model
 from dffml.util.entrypoint import entrypoint
 
-from .config import AutoSklearnConfig, AutoSklearnModelContext
+from .config import AutoSklearnConfig
+from .autosklearn_base import AutoSklearnModelContext, AutoSklearnModel
 
 
 class AutoSklearnClassifierModelContext(AutoSklearnModelContext):
     def __init__(self, parent):
         super().__init__(parent)
-
-    @property
-    def model(self):
-        """
-        Generates or loads a model
-        """
-        if self._model is not None:
-            return self._model
-        config = self.parent.config._asdict()
-        del config["predict"]
-        del config["features"]
-        del config["location"]
-        self._model = autosklearn.classification.AutoSklearnClassifier(
-            **config
-        )
-        return self._model
-
-    @model.setter
-    def model(self, model):
-        """
-        Loads a model if already trained previously
-        """
-        self._model = model
+        if self.parent.model is None:
+            config = self.parent.config._asdict()
+            del config["predict"]
+            del config["features"]
+            del config["location"]
+            self.parent.model = autosklearn.classification.AutoSklearnClassifier(
+                **config
+            )
 
     async def get_probabilities(self, data):
-        return self.model.predict_proba(data)
+        return self.parent.model.predict_proba(data)
 
 
 @entrypoint("autoclassifier")
-class AutoSklearnClassifierModel(Model):
+class AutoSklearnClassifierModel(AutoSklearnModel):
     CONFIG = AutoSklearnConfig
     CONTEXT = AutoSklearnClassifierModelContext
